Split out base usage from general usage
Base usage are for options that are available to all commands, whereas
the GLOBAL_USAGE options include some options that are general but no
longer global.  This change supports limiting the options to just the
basic options.

To have a command only support base usage, then replace $GLOBAL_USAGE
with i$BASE_USAGE in the help section, and put "BASE_USAGE" as the first
option spec in the call to `options` routine.