Genesis update (#451)
* Extracted generic Github class from Github Kit provider

This will be used to fetch genesis releases for updates

* Improve kit provider status output

* Split out base usage from general usage

Base usage are for options that are available to all commands, whereas
the GLOBAL_USAGE options include some options that are general but no
longer global.  This change supports limiting the options to just the
basic options.

To have a command only support base usage, then replace $GLOBAL_USAGE
with i$BASE_USAGE in the help section, and put "BASE_USAGE" as the first
option spec in the call to `options` routine.

* Fix getting non-production github release version

If user specified version by name, it should include draft and
prereleases when looking for it.  This was the initial intent, but the
wrong options were set when attempting it.

* Add `update` genesis command to get new version

Usage: genesis update [-c|--check] [-v|--version VERSION] [-d|--details] [-f|--force]

  -c, --check       Check if a newer version (or the specified version) is
                    available.

  -v, --version <x> Install the specified version instead of the latest

  -d, --details     Print the release notes of the versions between the current
                    version and the latest (or the release notes of the
                    specified version)

  -f, --force       Replace the current genesis executable.  Otherwise,
                    confirmation will be requested.

* Added release notes