Allow Genesis::Kit#metadata to take args
If Genesis::Kit#metadata recieves no arguments, it will return a hash
reference for all metadata for the kit.  If it gets a single value, it
returns the metadata corresponding to the key of that value.  Finally,
if it gets a list of values, it will return a list of the metadata
corresponding to the keys of the given values.