Provide create-env deployments with diff/confirm
Since BOSH does not provide a confirmation after a diff on the
environment to be deployed, Genesis provides this based on a previous
deployment, assuming it was added to the repo.

If the exodus data from the previous deployment doesn't match the local
archived manifest, a warning will be stated, or in the case of running
under `-y`, the deploy will be aborted.