Further fixes to CI repipe from Github extraction
When Genesis::Github was extracted from the Kit::Provider::Github, some
properties were extracted into the remote object that replaced the
internal representation of the Github repository (is_a changed to
has_a).

As such, delegation routines were added that provided access to these
internal properties, but anything trying to access the properties
directly would fail, as is the case with the CI::Legacy lookup for the
organization property that was fixed in this commit.