Allow run to interpolate env vars in arguments
By setting the {eval_var_args => 1} option, any environment variables in
the arguments will be interpolated.  It will honour \$ to escape
interpolation, and will also honour braces to limit scope.

Unknown variables will be rendered as empty string (this includes
special variables that are not environment variables)  If you want to
use other variables, embed them in the call:

run(
  {eval_var_args => 1},
  "echo Home: $1, process: $$",
  '$HOME'
)

In the previous example, the $HOME will be interpolated and provided at
$1, while the $$ will be rendered by the shell itself.