Make bail auto-label and wrap
Messages from bail command will now auto label with a red [ERROR] and
wrap the message at the output column boundry.  If it already has a
label, and the rest of the message is wrapped to indent under the label,
it will unwrap it and reindent it correctly for the output width.