WIP: Moved Service libraries out of Genesis namespace
Genesis was holding all the sublibraries, but for services such as
Vault, Github, Credhub and the BOSH director, it makes more sense to
decouple them from Genesis and put them in their own Services namespace.

This is commit #1 in the process, which is simply moving them -- the
next step is to resolve the use/require calls.