Don't delete inherited tempdir (+ cleanup)
Refactored Genesis::workdir so that if it inherited the directory from a
parent genesis process, it wouldn't delete it on exit of the current
process.  As a side effect, you can set and export GENESIS_WORKDIR
environmental variable before running genesis commands and have control
of where the genesis work will be placed, and it will stick around after
the command completes for debugging purposes.

Also cleaned up some misplaced and unneeded coded, and captured the
error code when creating a file ($! was used insted of $@ inside eval)