Refactored out Genesis::IO package
This package was a half-baked datafile support language that allowed
json data to be dumped to a file, and yaml to be dumped to a file,
loaded from a file and loaded from a string (via a temporary file).

Equivalent or better routines to do that and more are available in the
core Genesis package, so this has been removed and calls refactored to
use the equivalent routines from Genesis.