Makes genesis secrets-provider not need a vault
In order to facilitate resetting the secrets provider when there is an
error in the existing secrets provider, the initial state of this
command must assume there to not be a valid vault.  It will then probe
the current vault state, to determine if it's valid and can be provided
as a default value.

[Bug Fixes]
* `genesis secrets-provider` no longer gets into a chicken/egg issue
  when trying to fix an invalid secrets provider.