Allow secrets plan to validate subject uniqueness
Subject Distinguished Name (DN) is suppose to be globally unique, but in
practicality, they should at least be unique within the cert chain,
because the issuer DN is compared to the subject DN in the chain to
identify the signer of a given cert.  Genesis has always validated this
condition for vault-based kits, but once we started vaultifying
credhub-based kits (ie cf), we realized that cf doesn't follow the
rules, so exceptions had to be added to allow already published cf kits
to continue to work.