Make entombment and vaultification based on version
This change makes the decision to entomb and/or vaultify secrets based on
the version of the environment being deployed, rather than based on
command-line options.  This is a more consistent and predictable
behavior, and allows for better control over the behavior of Genesis
deployments across different users.

In order to enable bosh entombment and vaultification, the environment
must be at least version 3.0.0-rc.1.  To disable either of these
behaviors, the user can set `genesis.entomb` or `genesis.vaultify` to
`false` in their environment yml file.