Specify version of Genesis in new environment files
New environment creation now uses the current version of Genesis as the
minimum version if not specified.  This is to ensure that the new
environment is compatible with the current version of Genesis, and
benefits from any new features or bug fixes, while simultaneously
older versions can't use environments that may contain features or
behaviors that might have been used in the file.