Resolve "secret ... not found" errors in adaptive merge
[Bug Fixes]

- Fixes 'secret .. not found' errors when merging a manifest under certain
  conditions when trying to partially resolve a manifest without all the
  secrets or yaml fragments available.

---

When a secret is not found in the merged manifest, we should look in the
original content for the secret and replace it with a `(( defer vault
... ))` operation. This will allow the adaptive merge to proceed without
erroring out.

This condition is caused by the adaptive merge trying to resolve a
secret that was removed from the merged manifest by a prune operation
that was deferred as part of the adaptive merge (or similar).