Extract log styles
This commit extracts the log styles into a separate method, `log_styles`,
which returns a hashref of the log styles.  This allows the styles to be
reused in other parts of the codebase, such as in the `dump_var` and
`dump_stack` methods.

It also provides the `style` method to query the style of a given log.