Added validation to the configuration system
This commit adds a new `validate` method to the `Genesis::Config` class
which allows for the configuration to be validated against a schema. The
schema is a hashref that describes the expected configuration keys and
their types, and can also specify default values, required keys, and
allowed values for enums.

The `validate` method will ensure that all required keys are present,
and that all keys are of the correct type. It will also normalize the
values of boolean keys to JSON::PP::true or JSON::PP::false, and will
ensure that enum values are one of the allowed values.

This is initially being used to validate the user's configuration file
under `~/.genesis/config`, but it could be used in other places as well.