Changes the init command to drop the '-deployments' suffix
As of Genesis 3.0.0, the default directory name for a Genesis deployment
repository no longer has the '-deployments' suffix.  This change is
intended to make the directory name more concise and easier to work with
in the shell.

However, the '-d' option to the 'init' command is still available for
those who wish to have an arbrary directory name.  Furthermore, the
'legacy_repo_suffix' option in the Genesis configuration file can be set
to true to restore the old behavior.