Added default_bosh_target user config option
This commit adds a new user configuration option, `default_bosh_target`,
which can be set to either `ask`, `self`, or `parent`.  If set to `ask`,
the user will be prompted to choose between `self` and `parent` when
targeting a BOSH director that was deployed by another BOSH director
when using the `bosh` command if the --self or --parent options are not
provided.