Add `genesis environments` command to list env files and their status
[Improvements]

- Add `genesis environments` command to list all environments in the
  known deployment repositories.  This command will list the
  environments, their kit, and the last deployment status.  The
  environments can be grouped by the environment name or by the kit
  repository (default).

---

Adds ansi control sequences to hide and show the cursor when running
`genesis environments` to provide a more polished experience.  It also
ensures that the cursor is shown when the command completes.

It also enables silence mode for validating the vault connection when
running `genesis environments` to prevent the status from being printed
in the middle of processing the environments (each repo potentially has
a different vault connection).  It will still print the status if the
connection times out or fails.