Alter user config for warnings and deploy roots
Move the `allow_oversized_secrets` setting to a new `suppress_warnings`
hash, and rename it to `oversized_secrets`.  This aligns with the way we
suppress warnings for bosh_target.

Also, provide an environment variable for `GENESIS_DEPLOYMENT_ROOTS` to
allow users to set or override the deployment roots in their environment.