Extend the Genesis configuration schema
This change adds support for a new schema type, hash , which allows for
nested configuration values. This was needed to validate the existing
repository-based configuration files.  This also extends the validation to
support deep nesting (e.g. hash of hashes, array of hashes, etc.).

The schema validation now also supports a new key, envvar , which allows
for a configuration value to be set or overridden from an environment
variable.

Finally, this change adds a new schema type, semver , which validates
that a configuration value is a valid semantic version string.