Genesis::Top now validates the repo config file
Top, which represents the deployment repo, now validates the configuration
file when it is loaded. This will help prevent errors in the future when
we add new configuration options.

It also makes use of the recent change to Genesis::Config to allow for
environment variables to be used in the configuration file to override
values and to allow for default values to be set.