Enhanced config schema validation
This change adds the ability to specify multiple types for a single
schema key. This is useful when a key can be one of several types, such
a semver or a specific string.

Also added explicit string and null types, which are useful for
validating specific values.

Finally, added a 'any' type, which is a no-op. This is useful when
validating a key that can be any type, and you just want to validate
that it is a valid key.

Used the above to validate the creator_version key in the repo config,
which can be a semver or the string "(development)".