New Feature: genesis credhub
[Improvements]

This commit adds a new command, genesis credhub, which allows you to run
credhub commands against the CredHub server associated with the
environment without having to log in explicitly.  This command is similar
to the bosh command, in that it will automatically select the correct
CredHub server based on the environment type and the options provided.

As a convenience, the command will automatically augment the path or name
of the CredHub object based on the command being run.  For example, if you
run `genesis credhub get -n my-secret`, it will automatically prepend the
base path of the CredHub server to the name, so that it will be
`/<director-path>/<deployment-path>/my-secret`.  This behavior can be
bypassed with the `--raw` option.  This works for most functions that
take a name or path option, such as `get`, `set`, `delete`, `generate`,
`regenerate`, `export`, `interpolate`, and `find`.  Specifically
regarding the `find` command, if the `-p|--path` and the
`-n|--name-like` options are both provided, `find` will ignore the
`--path` option and use the `--name-like` option to search for any
matching name under the root path `/`.  This is the behaviour of the
credhub cli, and Genesis cannot override it.

---

This commit also moved the bosh target selection into a new method on the
Genesis::Env object, get_target_bosh, which will return the correct BOSH
director to target based on the environment type and the options provided.

The utility function index_of was added to Genesis.pm to find the index
of an item in an array.  This was used to augment the credhub command to
automatically provide the base name or path for the CredHub server based
on the command being run.  (the --raw option can be used to bypass this
behavior)