Finalized selector prompt for environment files
This commit finalizes the selector prompt for environment files.  It
adds the ability to display section headers in the list of choices, and
to display a different label for the selected item than the label
displayed in the list of choices.

This commit implements section breaks for prompt_for_choice, so the list
of choices and list of labels are no longer required to be the same
length.  If the label is a string that matches the pattern "---(.*)---",
it will be treated as a section header, and will be displayed as such,
without a selection number.  A label of "---" will be treated as a
section break without a header.

Decided to continue to allow unlabeled deployment roots, and validate
that labeled deployment roots do not start with an '@' character (reserved
for internal use).