Further refinements to the `genesis edit` command
This commit adds support for the `code` editor (VSCode) to the `genesis
edit` command.  It also adds a note to the output when multiple files
are opened in VSCode, as they are opened in separate tabs by default.

Additionally, this commit refactors the `genesis edit` command to
display the full path to the environment file being edited, rather than
just the relative path.

Also fixed the display of the ambiguous environment file error message
when not in a controlling terminal -- the section headers were being
displayed as the file labels, which was confusing and just plain wrong.