Revert humanize_path to not use label_maps
This reverts commit 3b1b3b1b, which added the ability to specify label
maps to give context to the root path of a given path.  This behavior
was considered too confusing and not useful enough remain.

Instead, we replace the label_maps argument with an absolute argument,
which will force the path to be displayed as an absolute path, rather
than a relative path, as a bunch of ../ in the path can be confusing.