Added ability to call some internal methods from cli
This is a special command that is not intended to be used by end users,
but rather by developers of Genesis itself.  It allows for running a
method with arguments inside the Genesis environment, which is useful for
testing and development.

For example, to run the `Genesis::Env->search_for_env_file` method with
the arguments 'dev' and 'default', you would run:

  genesis hack Genesis::Env->search_for_env_file dev default