Fixes a bug in the markdown table rendering code
The markdown table rendering code was not correctly calculating the
minimum width of each column when the min and max widths were the same.

Also, fixed a bug where old-style link references were not being
converted to the new-style link references.

Finally, fix a logic bug in the assessing the min widths of each column.