Update `genesis envs` use deployment_roots_map
This change updates the `genesis envs` command to use the new
`Genesis::deployment_roots_map` function to get the list of deployment
roots.

This change also fixes a bug where the `genesis envs` command would
fail if the environment had an invalid kit. Now, the command will
attempt to load the environment and if it fails, it will report the
invalid kit as "not found".