Allow `genesis edit` to use local kit replacement
This change allows `genesis edit` to prompt the user to select a local
kit to replace the kit specified in the environment file if the kit
specified in the environment file is not found. This allows the user to
continue editing the environment file with the manual of a local kit
version if the specified kit is not found.  If the editor is vi-based,
it will also alter the file to use the selected kit version.