Improve detection of Genesis repositories
Since $HOME can contain a .genesis directory that contains a config
file, we need to be more careful about how we detect Genesis
repositories.  This change adds a new method to Genesis::Top that
checks for the presence of a .genesis/config file and verifies that it
contains a deployment_type: key.  This should be a more reliable
indicator that a directory is a Genesis repository.