Allow top to be used without a vault
This change allows the `top` command to be used without a vault.  While
explicitly setting the `no_vault` option to true will allow the command
to run without a vault already exists for commands that do not require
a vault, we needed to add a way to allow a `top` object to be created
if the target vault was not defined, unsealed, unreachable, or otherwise
not available.

Specifically, the `genesis environments` command uses this mode to allow
it to show information in exodus if the vault is available, but handle
the case where the vault is not available gracefully.