Allow search prefix to be used with `genesis environments` command
This change allows the `genesis environments` command to be used with the
search prefix. This allows users to filter the environments that are
displayed by the command.  The only difference between this and normal
use of the search prefix is that if the deployment type is not
specified, all deployment types are included in the search, not just
`bosh` deployments.