Extract the script wrapper to new function fake_tty
This commit extracts the script wrapper logic from show_diff into a new
function called fake_tty. This function takes a filename and a list of
command arguments and returns a list of arguments that will run the
command in a script session that fakes being in a tty terminal.