Added IP4::Address, IP4::Range, and IP4::MultiRange modules.
These modules are used to represent and manipulate IPv4 addresses,
ranges, and collections of ranges.  The IP4::Address module can be used
to represent an IPv4 address and provides methods for converting between
integer and string representations, as well as methods for comparing
addresses.  The IP4::Range module can be used to represent a range of
IPv4 addresses and provides methods for converting between CIDR
notation, a list of addresses, and a list of ranges.  The
IP4::MultiRange module can be used to represent a collection of
IP4::Range objects and provides methods for set union, intersection, and
exclusion.  The IP4::MultiRange module also provides methods for
compacting a collection of ranges by combining overlapping or contiguous
ranges.

Fixed bugs, flushed out IP4 library

* Added subtract method to IP4::Address
* Added to method to IP4::Address to form a range

* Added slice method to IP4::Range
* Added slice method to IP4::MultiRange

* Added range method to IP4::MultiRange
* Added subtract method to IP4::MultiRange

* Extended initializer for IP4::Range to accept an array of two IP4::Address objects or two strings
* Extended initializer for IP4::MultiRange to accept an array of IP4::Range objects or IP4::MultiRange objects

* Fixed various bugs in IP4::Range and IP4::MultiRange found during testing