WIP: Spike on refactoring kit providers
This allows multiple kit providers can me specified in a hierarchical
order, some are allowed to be sources (genesis-community, github, repo,
vault, filesystem-archive), some are only allowed to be destinations (dev, local),
some can be used for deployments (vault, repo, local, dev).  The
configuration for these for order and applicability is stored in the
repo configuration file, which will be settable by the `genesis init
--kit-providers-config path/to/source.yml` option, and later by the
`genesis config` option various ways

This initial commit contains most of the config changes and validation,
along with some initial changes to the kit provider modules and the kit,
env and top modules to support the new provider types.