Added autoloading of secrets when accessing value
Off by default, this change adds an `autoload` method to the `Genesis::Secret`
class that will load the secret if it hasn't been loaded yet. This is useful
for secrets that are not needed in all cases, but are needed in some cases.

The Secret::Plan class has been updated to add an autoload method that will
turn on autoloading for all secrets in the plan as a convenience.