[WIP] Add bosh-configs command and refactor bosh commands
This commit adds a new `bosh-configs` command to Genesis, which will
allow users to create, list, view, compare, and delete bosh configs
for an environment.  This command will be used to manage the various
bosh configs that are uploaded to the BOSH director for an environment.

This commit also refactors the `bosh`, `credhub`, `logs`, and `broadcast`
commands to be part of the `Genesis::Commands::Bosh` module, rather than
`Genesis::Commands::Env`.  This is to better organize the commands and
to make it easier to add new commands in the future.