Added ability to create a config with explicit content
This allows for the creation of a config object with a specific set of
content, rather than having to load it from a file. This is useful for
initiating a config object with a set of defaults, or handling sub-configs
that are pulled out of a larger config.

This change also includes a small change to the autosave logic, which was
previously only checking if the autosave was set, rather than if the path
was defined and the autosave flag was set. (Can't autosave if you don't
have a path to save to, but this wasn't a problem when configs could only
be loaded from files.)