Enhance flatten/unflatten to handle keys with dots in them
This change allows the flatten/unflatten functions to handle keys that
contain dots.  This is done by replacing dots in the key with a tilde
(~) character when flattening, and then replacing the tilde with a dot
when unflattening.

Also included is a diminutive form of the flatten function that can be
called with a single argument, which is the value to be flattened.  This
is useful for calling the function for the first time.