Add set_path and get_path to Service::Vault
This commit adds two new methods to the Service::Vault module, `set_path`
and `get_path`. These methods allow for the setting and getting of multiple
key value pairs in a single call, and are useful for managing large sets of
secrets in the vault.

It also supports automatic flattening of the data structure, so that nested
hashes can be written to the vault as a single key, and then read back as
a nested hash.