Change manifest store values and default
This commit changes the default value for the `manifest_store` key in the
top-level configuration file to `hybrid`, and changes `repo` to `repository`
as a valid value.

The `hybrid` value is a new value that will be used to transition from the
legacy `repository` storage mechanism to the new `exodus` storage mechanism,
while staying backwards compatible with previous versions of Genesis.
Once firmly established and comfortable with the deployments stored in
exodus, the `manifest_store` value can be changed to `exodus` by the
operator.