Support search prefix type in GENESIS_CALL
This commit adds support for the `search` prefix type in the `GENESIS_CALL`
environment variable. This allows the hooks to reference the genesis call that
was used to run the command so that it can inform the user to use the
same style of call to run other subsequent commands.