Use bosh connection check differently for BOSH_ALL_PROXY
When using BOSH_ALL_PROXY, we can't rely on the tcp_listening check to
determine if the BOSH director is reachable. Instead, we need to run the
`bosh env` command to determine if the director is reachable, and rely
on a timeout to prevent the command from hanging indefinitely.

While this would work for the normal case as well, tcp_listening is much
faster at resolving the connection status.