Refactor director_exodus_lookup
This commit refactors the director_exodus_lookup method to use the new
bosh object and its vault method to retrieve the exodus data for the
director from the vault independently of the environment's vault.  It
also supports extended exodus paths, which allow for the retrieval of
exodus data from a sub-path of the director's exodus data.