Major Hooks Overhaul, including new cloud-config
Added a new `cloud-config` hook, with support for a purpose-driven
`director` sub-type.  This is used to generate and upload per-kit
cloud-configs with names associated with just the single deployment,
preventing name collisions in cloud configs.  The bosh director also
specifies azs, and compilation sections as well as the regular networks,
vm_types, disk_types and optional vm_extensions sections that nominal cloud
configs have.

The cloud-config hook is run before the deploy, and the generated cloud
config is compared to the existing cloud config on the director.  If there
are differences, the user is prompted to upload the new cloud config to the
director.  If the user declines, the deploy is aborted.

The hook file is expected to return a hashref with two keys: `cloud_config`
which is a string containing the cloud config, and `network_map` which is
a hashref containing the network map for the deployment.  The network map
is used to update the director's network data in the exodus vault path.
Currently only Perl-based cloud-config hooks are supported.

Also added Perl-based hook support for `features` and `post-deploy` hooks.

The `deploy` command in `Genesis::Commands::Env` has been refactored to
support the new cloud-config hook.  The `deploy` subroutine in `Genesis::Env`
has also been massively rewritten to better support the new cloud-config
as well as the outstanding changes to support multiple manifest stores.

The network map is passed in to the deploy method, and if the deploy is
successful, the network map is updated in the exodus vault path for the
director.

Future refactoring will need to address the fact that the new cloud
config doesn't get uploaded when just generating the manifest or running
checks.  This will be addressed in a future commit, most likely by
augmenting the cloud configs downloaded from the director with the new
generated cloud config.

It was originally intended to have a bosh-configs command to manage
these new features, but it was better served to have them integrated
directly into the deploy command, so they did not get out of sync with
any changes not yet deployed.