Improved create-env detection for 3.1.0-rc1
This commit improves the detection of create-env environments for
Genesis 3.1.0-rc1.  The previous implementation had some blind spots
that could lead to false positives or negatives.  This commit
introduces a new feature compatibility check for 3.1.0-rc1, and
introduces a simplified logic for determining if an environment is a
create-env environment.  It also bypasses the check proto in features
because this was causing a cycle when features was checking for create-env.
(Caveat: the feature hook can no longer inject the proto feature into the
features list, but this is a minor issue as proto has been deprecated)