Fix requirement for iaas and scale
Genesis 3.1.x introduced a new feature where the `iaas` and `scale`
values are required to be set in the `bosh-configs` section of the
environment file.  However, this is not needed for the rest of the kit
hooks, and was causing issues with kits written for Genesis 3.0.x and
previous versions.

This commit fixes the requirement for iaas and scale to be declared in
the environment, only requiring it if the kit requires it or in the case
of iaas, if it is a create-env environment.

Kits can set `requires_iaas` and `requires_scale` in their `kit.yml` to
true to require these values to be set in the environment.

A future commit may change the type of these values to a list of hooks
instead of a boolean, to allow for more fine-grained control over which
hooks require these values.