Optionally make iaas and scale (non-)fatal errors
Genesis 3.1.x introduced a new feature where the `iaas` and `scale`
values are required to be set in the `bosh-configs` section of the
environment file.  However, this is not needed for the rest of the kit
hooks, and was causing issues with kits written for Genesis 3.0.x and
previous versions.

This change makes the `iaas` and `scale` methods optionally fatal, so
that they are only required to be set when the `cloud-config`, `manifest`
or `deploy` hooks are being run.