Synthesize a pruned manifest from unpruned data
If a manifest is requested to be pruned, but is only available in an
unpruned form, we can now synthesize a pruned version of the manifest
from the unpruned data. This is done by first checking that the unpruned
data is not tampered with, then pruning the data, and finally generating
a new sha1 hash for the pruned data.